+++ embassy-stm32/src/usart/v2.rs
@@ -87,7 +87,7 @@ impl<'d, T: Instance, TxDma, RxDma> Uart<'d, T, TxDma, RxDma> {
             });
         }
         let r = self.inner.regs();
-        let dst = r.tdr().ptr() as *mut u32;
+        let dst = r.tdr().ptr() as *mut u8;
         crate::dma::write(ch, request, buffer, dst).await;
         Ok(())
     }
@@ -104,7 +104,7 @@ impl<'d, T: Instance, TxDma, RxDma> Uart<'d, T, TxDma, RxDma> {
             });
         }
         let r = self.inner.regs();
-        let src = r.rdr().ptr() as *mut u32;
+        let src = r.rdr().ptr() as *mut u8;
 
         crate::dma::read(ch, request, src, buffer).await;
         Ok(())
